app/fetch/requirements.txt
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
click==8.0.1
fastapi==0.68.0
h11==0.12.0
idna==3.2
multidict==5.1.0
numpy==1.21.1
pandas==1.3.1
pydantic==1.8.2
PyJWT==2.1.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
yarl==1.6.3